Refactor gatsby-{browser,ssr}.tsx by adding gatsby-shared.tsx
